src/core/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = [
  "poetry-core",
]

[tool.poetry]
name = "trulens-core"
version = "1.2.6"
description = "Library to systematically track and evaluate LLM based applications."
authors = [
  "Snowflake Inc. <ml-observability-wg-dl@snowflake.com>",
]
license = "MIT"
readme = "README.md"
packages = [
  { include = "trulens" },
]
homepage = "https://trulens.org/"
documentation = "https://trulens.org/getting_started/"
repository = "https://github.com/truera/trulens"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
  "Development Status :: 3 - Alpha",
  "License :: OSI Approved :: MIT License",
]

[tool.poetry.dependencies]
python = "^3.8.1"
numpy = [
  { version = ">=1.23.0,<1.25.0", python = "<3.9" },
  { version = ">=1.23.0", python = ">=3.9" },
]
munch = "^2.5"
dill = ">=0.3.8"
typing_extensions = "^4.9"
pydantic = "^2.4.2"
pandas = ">=1.0.0"
rich = "^13.6"
sqlalchemy = "^2.0"
alembic = "^1.8.1"
nest-asyncio = "^1.5"
python-dotenv = ">=0.21,<2.0"
importlib-resources = "^6.0"
trulens-semconv = { version = "^1.0.0", optional = true }
opentelemetry-api = { version = "^1.0.0", optional = true }
opentelemetry-sdk = { version = "^1.0.0", optional = true }

[tool.poetry.group.tqdm]
optional = true

[tool.poetry.group.tqdm.dependencies]
tqdm = { version = ">=4.2.0", optional = true }

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
trulens-semconv = { path = "../semconv" }
